+++ aiohttp_swagger3/__init__.py
@@ -1,5 +1,5 @@
 __all__ = ("SwaggerDocs", "SwaggerFile", "__version__")
-__version__ = "0.1.8"
+__version__ = "0.2.0"
 __author__ = "Valetov Konstantin"
 
 from .swagger_docs import SwaggerDocs